requirements.txt
Django>=4.0
channels>=4.0.0
daphne>=4.0.0
twisted[http2,tls]>=22.0.0
psycopg2-binary>=2.9.0
google-generativeai>=0.3.0
Pillow>=9.0.0
django-ckeditor>=6.5.0
python-pptx>=0.6.21
beautifulsoup4>=4.11.0
requests>=2.28.0
python-docx>=0.8.11
PyPDF2>=2.11.0
langchain>=0.0.267
langchain-community>=0.0.1
python-magic>=0.4.27
jinja2>=3.1.0
lxml>=4.9.0
markdown>=3.0.0
pygments>=2.0.0
pyquery>=1.0.0
pyyaml>=6.0.0
whitenoise>=6.0.0